chore: ProfileDeleteAccountButton added among others
